third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_continue_objectstore2-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.continue() - object store - attempt to pass a key parameter is not a valid key  
Harness: the test ran to completion.
